feature: 报告模块接口梳理及逻辑修改完成

+++ app-monitor/src/main/java/com/dataint/monitor/dao/IReportDao.java
@@ -11,7 +11,7 @@ import java.util.Date;
 import java.util.List;
 
 @Repository
-public interface IReportDao extends JpaRepository<Report, Integer> {
+public interface IReportDao extends JpaRepository<Report, Long> {
 
     Page<Report> findByGmtStartAndGmtEndAndReportType(Date gmtStart, Date gmtEnd, String reportType, Pageable pageable);
